refactor(store): extract build and unlock actions
