+++ src/models/store/actions/buildTower.ts
@@ -1,0 +1,107 @@
+import type { Store } from '../index';
+import type { Tower } from '../../gameTypes';
+import { GAME_CONSTANTS } from '../../../utils/constants';
+import { upgradeEffectsManager } from '../../../game-systems/UpgradeEffects';
+import { energyManager } from '../../../game-systems/EnergyManager';
+import { useGameStore } from '../index';
+
+export function buildTowerAction(
+  state: Store,
+  slotIdx: number,
+  free = false,
+  towerType: 'attack' | 'economy' = 'attack',
+  towerClass?: string
+): Partial<Store> {
+  const slot = state.towerSlots[slotIdx];
+  if (!slot || slot.tower) return {};
+
+  const cost = free ? 0 : GAME_CONSTANTS.TOWER_COST;
+  if (!free && state.gold < cost) return {};
+
+  const upgrade = GAME_CONSTANTS.TOWER_UPGRADES[0];
+
+  let specializedTowerData = null as any;
+  if (towerClass && GAME_CONSTANTS.SPECIALIZED_TOWERS[towerClass as keyof typeof GAME_CONSTANTS.SPECIALIZED_TOWERS]) {
+    specializedTowerData = GAME_CONSTANTS.SPECIALIZED_TOWERS[towerClass as keyof typeof GAME_CONSTANTS.SPECIALIZED_TOWERS];
+    if (!free && state.gold < specializedTowerData.cost) return {};
+  }
+
+  const finalCost = free ? 0 : (specializedTowerData?.cost || cost);
+
+  const newTower: Tower = {
+    id: `${Date.now()}-${Math.random()}`,
+    position: { x: slot.x, y: slot.y },
+    size: GAME_CONSTANTS.TOWER_SIZE,
+    isActive: true,
+    level: 1,
+    range: towerType === 'economy' ? 0 : (specializedTowerData?.baseRange || GAME_CONSTANTS.TOWER_RANGE),
+    damage: towerType === 'economy' ? 0 : (specializedTowerData?.baseDamage || upgrade.damage),
+    fireRate: towerType === 'economy' ? 0 : (specializedTowerData?.baseFireRate || upgrade.fireRate),
+    lastFired: 0,
+    health: upgrade.health,
+    maxHealth: upgrade.health,
+    wallStrength: upgradeEffectsManager.applyShieldUpgrades(state.globalWallStrength),
+    specialAbility: upgrade.special,
+    healthRegenRate: 0,
+    lastHealthRegen: 0,
+    specialCooldown: 5000,
+    lastSpecialUse: 0,
+    multiShotCount: 3,
+    chainLightningJumps: 5,
+    freezeDuration: 2000,
+    burnDuration: 3000,
+    acidStack: 0,
+    quantumState: false,
+    nanoSwarmCount: 5,
+    psiRange: 150,
+    timeWarpSlow: 0.3,
+    spaceGravity: 0.5,
+    legendaryAura: false,
+    divineProtection: false,
+    cosmicEnergy: 100,
+    infinityLoop: false,
+    godModeActive: false,
+    attackSound: towerType === 'economy' ? undefined : GAME_CONSTANTS.TOWER_ATTACK_SOUNDS[0],
+    visual: GAME_CONSTANTS.TOWER_VISUALS.find(v => v.level === 1),
+    rangeMultiplier: slot.modifier?.type === 'buff' ? GAME_CONSTANTS.BUFF_RANGE_MULTIPLIER : 1,
+    towerType,
+    towerCategory: specializedTowerData?.category,
+    towerClass: towerClass as any,
+    criticalChance: specializedTowerData?.criticalChance ?? 0,
+    criticalDamage: specializedTowerData?.criticalDamage ?? 1,
+    armorPenetration: specializedTowerData?.armorPenetration ?? 0,
+    areaOfEffect: specializedTowerData?.areaOfEffect ?? 0,
+    projectilePenetration: specializedTowerData?.projectilePenetration ?? 0,
+    spinUpLevel: specializedTowerData?.spinUpLevel ?? 0,
+    maxSpinUpLevel: specializedTowerData?.maxSpinUpLevel ?? 0,
+    beamFocusMultiplier: specializedTowerData?.beamFocusMultiplier ?? 1,
+    beamLockTime: specializedTowerData?.beamLockTime ?? 0,
+    supportRadius: specializedTowerData?.supportRadius ?? 0,
+    supportIntensity: specializedTowerData?.supportIntensity ?? 1,
+    shieldStrength: specializedTowerData?.shieldStrength ?? 0,
+    shieldRegenRate: specializedTowerData?.shieldRegenRate ?? 0,
+    repairRate: specializedTowerData?.repairRate ?? 0,
+    empDuration: specializedTowerData?.empDuration ?? 0,
+    stealthDetectionRange: specializedTowerData?.stealthDetectionRange ?? 0,
+    manualTargeting: false,
+    upgradePath: '',
+    synergyBonuses: { damage: 0, range: 0, fireRate: 0 }
+  };
+
+  const newSlots = [...state.towerSlots];
+  newSlots[slotIdx] = { ...slot, tower: newTower, wasDestroyed: false };
+  state.towerUpgradeListeners.forEach(fn => fn(newTower, 0, 1));
+
+  setTimeout(() => {
+    import('../../../utils/sound').then(({ playContextualSound }) => {
+      playContextualSound('tower-build');
+    });
+  }, 50);
+
+  return {
+    towerSlots: newSlots,
+    towers: [...state.towers, newTower],
+    gold: state.gold - finalCost,
+    totalGoldSpent: state.totalGoldSpent + finalCost,
+  };
+}
